ext: adjust index type
* ext: use long for index instead of int and RARRAY_LENINT.

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@52447 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ext/fiddle/function.c
@@ -90,7 +90,7 @@ initialize(int argc, VALUE argv[], VALUE self)
     ffi_type **arg_types;
     ffi_status result;
     VALUE ptr, args, ret_type, abi, kwds;
-    int i;
+    long i;
 
     rb_scan_args(argc, argv, "31:", &ptr, &args, &ret_type, &abi, &kwds);
     if(NIL_P(abi)) abi = INT2NUM(FFI_DEFAULT_ABI);
